// seynekun/ClientApp/src/app/seynekun/models/modelo-lote/lote.ts
export class Lote {
    numeroLote: number;
    variedad: string;
    numeroArbol: number;
    sistemaRenovacion: string;
    fechaSiembra: Date;
    cultivo: string;    
    fechaCosecha: Date;
    fechaFloracion: Date;
    estimadoCosecha: number;
    tipoEstimado: string;
    distanciaSiembra: number;
}